test: delete tests that don't test anything
Pure-noise tests removed:
- TenantIdTests, TenantHostIdTests — value-object equality is guaranteed
  by the C# record struct compiler.
- ExceptionTests — 16 tests of constructor passthrough and string
  interpolation; no behavior under test.
- IModuleTests.DefaultMethods_DoNotThrow — interface default-method
  invocation cannot fail.
- RateLimitingModuleTests — asserts hardcoded compile-time policy
  config; refactor-and-rename tests with no domain logic.
- DevTools/FileWatcherIntegrationTests — three smoke tests that start a
  file watcher and assert no exception. Never verified the watcher
  actually noticed the file change.
- OpenIddict ConnectEndpointTests — tests that OpenIddict middleware
  rejects bad requests; framework behavior, not module logic. The
  AreRegistered test was an explicit not-404 smoke check.
- Admin GetUsersEdit_ExistingUser_Returns200 — was Skipped due to test
  infrastructure gap; the live _NonExistent_Returns404 sibling already
  covers the routing.

Extracted FakeHostEnvironment to its own file (was used by
ServiceLifecycleTests too). ~30 tests deleted, no coverage lost.
